Rename simul to simulation
